requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn[standard]
python-multipart
pydantic